\name{rmsap-package}
\alias{rmsap-package}

\docType{package}
\title{
Statistical Analyses for Methylation Sensitive Amplification Polymorphism assays
}
\description{
This package provides tools for Statistical Analyses of Methylation Sensitive Amplification Polymorphism (MSAP) assays
}
\details{
\tabular{ll}{
Package: \tab rmsap\cr
Type: \tab Package\cr
Version: \tab 0.01\cr
Date: \tab 2012-05-08\cr
License: \tab GPL-2\cr
Requires: \tab ade4, pegas\cr
LazyLoad: \tab yes\cr
}


}
\author{
Andres Perez-Figueroa (\email{anpefi@uvigo.es})
}

