\name{polymorphic}
\alias{polymorphic}

\title{
	Checks polymorphic state
}
\description{
	Checks if a list of binary aleles is polymorphic (with at least two ocurrences of both states)
}
\usage{
polymorphic <- function(col)

}
\arguments{
  	\item{col}{
  		A list of binary alleles
	}
      
}

\author{
	Andres Perez-Figueroa (\email{anpefi@uvigo.es})
}

