\name{shannon}

\title{
Shannon's Diversity Index
}
\description{
Calculate Shannon's Diversity Index given a list of binomial alleles.   
}
\usage{
shannon(x)

}
\arguments{
  \item{x}{
  A list of binomial values or alleles. 
}
  
}
\details{
  



  
}
\value{
	The value of the Shannon's I
}

\note{
  
}
\references{

}
\author{
Andres Perez-Figueroa (\email{anpefi@uvigo.es})
}
\seealso{
	
}
\examples{
al <- c(1,1,0,0,0,0,1,0,0,0,1,1)
shannon(al)
}
